update to 1.5

+++ es_extended/client/main.lua
@@ -1,11 +1,12 @@
 local pickups = {}
 
-Citizen.CreateThread(function()
+CreateThread(function()
 	while not Config.Multichar do
-		Citizen.Wait(0)
+		Wait(0)
 		if NetworkIsPlayerActive(PlayerId()) then
 			exports.spawnmanager:setAutoSpawn(false)
 			DoScreenFadeOut(0)
+			Wait(500)
 			TriggerServerEvent('esx:onPlayerJoined')
 			break
 		end
@@ -20,14 +21,14 @@ AddEventHandler('esx:playerLoaded', function(xPlayer, isNew, skin)
 	FreezeEntityPosition(PlayerPedId(), true)
 
 	if Config.Multichar then
-		Citizen.Wait(3000)
+		Wait(3000)
 	else
 		exports.spawnmanager:spawnPlayer({
 			x = ESX.PlayerData.coords.x,
 			y = ESX.PlayerData.coords.y,
 			z = ESX.PlayerData.coords.z + 0.25,
 			heading = ESX.PlayerData.coords.heading,
-			model = `mp_m_freemode_01`,
+			model = GetHashKey("mp_m_freemode_01"),
 			skipFade = false
 		}, function()
 			TriggerServerEvent('esx:onPlayerSpawn')
@@ -47,7 +48,7 @@ AddEventHandler('esx:playerLoaded', function(xPlayer, isNew, skin)
 		end)
 	end
 
-	while ESX.PlayerData.ped == nil do Citizen.Wait(20) end
+	while ESX.PlayerData.ped == nil do Wait(20) end
 	-- enable PVP
 	if Config.EnablePVP then
 		SetCanAttackFriendly(ESX.PlayerData.ped, true, false)
@@ -99,33 +100,36 @@ end)
 
 AddEventHandler('skinchanger:modelLoaded', function()
 	while not ESX.PlayerLoaded do
-		Citizen.Wait(100)
+		Wait(100)
 	end
 	TriggerEvent('esx:restoreLoadout')
 end)
 
 AddEventHandler('esx:restoreLoadout', function()
 	ESX.SetPlayerData('ped', PlayerPedId())
-	local ammoTypes = {}
-	RemoveAllPedWeapons(ESX.PlayerData.ped, true)
 
-	for k,v in ipairs(ESX.PlayerData.loadout) do
-		local weaponName = v.name
-		local weaponHash = GetHashKey(weaponName)
+	if not Config.OxInventory then
+		local ammoTypes = {}
+		RemoveAllPedWeapons(ESX.PlayerData.ped, true)
 
-		GiveWeaponToPed(ESX.PlayerData.ped, weaponHash, 0, false, false)
-		SetPedWeaponTintIndex(ESX.PlayerData.ped, weaponHash, v.tintIndex)
+		for k,v in ipairs(ESX.PlayerData.loadout) do
+			local weaponName = v.name
+			local weaponHash = GetHashKey(weaponName)
 
-		local ammoType = GetPedAmmoTypeFromWeapon(ESX.PlayerData.ped, weaponHash)
+			GiveWeaponToPed(ESX.PlayerData.ped, weaponHash, 0, false, false)
+			SetPedWeaponTintIndex(ESX.PlayerData.ped, weaponHash, v.tintIndex)
 
-		for k2,v2 in ipairs(v.components) do
-			local componentHash = ESX.GetWeaponComponent(weaponName, v2).hash
-			GiveWeaponComponentToPed(ESX.PlayerData.ped, weaponHash, componentHash)
-		end
+			local ammoType = GetPedAmmoTypeFromWeapon(ESX.PlayerData.ped, weaponHash)
 
-		if not ammoTypes[ammoType] then
-			AddAmmoToPed(ESX.PlayerData.ped, weaponHash, v.ammo)
-			ammoTypes[ammoType] = true
+			for k2,v2 in ipairs(v.components) do
+				local componentHash = ESX.GetWeaponComponent(weaponName, v2).hash
+				GiveWeaponComponentToPed(ESX.PlayerData.ped, weaponHash, componentHash)
+			end
+
+			if not ammoTypes[ammoType] then
+				AddAmmoToPed(ESX.PlayerData.ped, weaponHash, v.ammo)
+				ammoTypes[ammoType] = true
+			end
 		end
 	end
 end)
@@ -138,6 +142,7 @@ AddEventHandler('esx:setAccountMoney', function(account)
 			break
 		end
 	end
+
 	ESX.SetPlayerData('accounts', ESX.PlayerData.accounts)
 
 	if Config.EnableHud then
@@ -147,77 +152,79 @@ AddEventHandler('esx:setAccountMoney', function(account)
 	end
 end)
 
-RegisterNetEvent('esx:addInventoryItem')
-AddEventHandler('esx:addInventoryItem', function(item, count, showNotification)
-	for k,v in ipairs(ESX.PlayerData.inventory) do
-		if v.name == item then
-			ESX.UI.ShowInventoryItemNotification(true, v.label, count - v.count)
-			ESX.PlayerData.inventory[k].count = count
-			break
+if not Config.OxInventory then
+	RegisterNetEvent('esx:addInventoryItem')
+	AddEventHandler('esx:addInventoryItem', function(item, count, showNotification)
+		for k,v in ipairs(ESX.PlayerData.inventory) do
+			if v.name == item then
+				ESX.UI.ShowInventoryItemNotification(true, v.label, count - v.count)
+				ESX.PlayerData.inventory[k].count = count
+				break
+			end
 		end
-	end
 
-	if showNotification then
-		ESX.UI.ShowInventoryItemNotification(true, item, count)
-	end
-
-	if ESX.UI.Menu.IsOpen('default', 'es_extended', 'inventory') then
-		ESX.ShowInventory()
-	end
-end)
-
-RegisterNetEvent('esx:removeInventoryItem')
-AddEventHandler('esx:removeInventoryItem', function(item, count, showNotification)
-	for k,v in ipairs(ESX.PlayerData.inventory) do
-		if v.name == item then
-			ESX.UI.ShowInventoryItemNotification(false, v.label, v.count - count)
-			ESX.PlayerData.inventory[k].count = count
-			break
+		if showNotification then
+			ESX.UI.ShowInventoryItemNotification(true, item, count)
 		end
-	end
 
-	if showNotification then
-		ESX.UI.ShowInventoryItemNotification(false, item, count)
-	end
+		if ESX.UI.Menu.IsOpen('default', 'es_extended', 'inventory') then
+			ESX.ShowInventory()
+		end
+	end)
 
-	if ESX.UI.Menu.IsOpen('default', 'es_extended', 'inventory') then
-		ESX.ShowInventory()
-	end
-end)
+	RegisterNetEvent('esx:removeInventoryItem')
+	AddEventHandler('esx:removeInventoryItem', function(item, count, showNotification)
+		for k,v in ipairs(ESX.PlayerData.inventory) do
+			if v.name == item then
+				ESX.UI.ShowInventoryItemNotification(false, v.label, v.count - count)
+				ESX.PlayerData.inventory[k].count = count
+				break
+			end
+		end
 
-RegisterNetEvent('esx:addWeapon')
-AddEventHandler('esx:addWeapon', function(weapon, ammo)
-	GiveWeaponToPed(ESX.PlayerData.ped, GetHashKey(weapon), ammo, false, false)
-end)
+		if showNotification then
+			ESX.UI.ShowInventoryItemNotification(false, item, count)
+		end
 
-RegisterNetEvent('esx:addWeaponComponent')
-AddEventHandler('esx:addWeaponComponent', function(weapon, weaponComponent)
-	local componentHash = ESX.GetWeaponComponent(weapon, weaponComponent).hash
-	GiveWeaponComponentToPed(ESX.PlayerData.ped, GetHashKey(weapon), componentHash)
-end)
+		if ESX.UI.Menu.IsOpen('default', 'es_extended', 'inventory') then
+			ESX.ShowInventory()
+		end
+	end)
 
-RegisterNetEvent('esx:setWeaponAmmo')
-AddEventHandler('esx:setWeaponAmmo', function(weapon, weaponAmmo)
-	SetPedAmmo(ESX.PlayerData.ped, GetHashKey(weapon), weaponAmmo)
-end)
+	RegisterNetEvent('esx:addWeapon')
+	AddEventHandler('esx:addWeapon', function(weapon, ammo)
+		GiveWeaponToPed(ESX.PlayerData.ped, GetHashKey(weapon), ammo, false, false)
+	end)
 
-RegisterNetEvent('esx:setWeaponTint')
-AddEventHandler('esx:setWeaponTint', function(weapon, weaponTintIndex)
-	SetPedWeaponTintIndex(ESX.PlayerData.ped, GetHashKey(weapon), weaponTintIndex)
-end)
+	RegisterNetEvent('esx:addWeaponComponent')
+	AddEventHandler('esx:addWeaponComponent', function(weapon, weaponComponent)
+		local componentHash = ESX.GetWeaponComponent(weapon, weaponComponent).hash
+		GiveWeaponComponentToPed(ESX.PlayerData.ped, GetHashKey(weapon), componentHash)
+	end)
 
-RegisterNetEvent('esx:removeWeapon')
-AddEventHandler('esx:removeWeapon', function(weapon)
-	local playerPed = ESX.PlayerData.ped
-	RemoveWeaponFromPed(ESX.PlayerData.ped, GetHashKey(weapon))
-	SetPedAmmo(ESX.PlayerData.ped, GetHashKey(weapon), 0)
-end)
+	RegisterNetEvent('esx:setWeaponAmmo')
+	AddEventHandler('esx:setWeaponAmmo', function(weapon, weaponAmmo)
+		SetPedAmmo(ESX.PlayerData.ped, GetHashKey(weapon), weaponAmmo)
+	end)
 
-RegisterNetEvent('esx:removeWeaponComponent')
-AddEventHandler('esx:removeWeaponComponent', function(weapon, weaponComponent)
-	local componentHash = ESX.GetWeaponComponent(weapon, weaponComponent).hash
-	RemoveWeaponComponentFromPed(ESX.PlayerData.ped, GetHashKey(weapon), componentHash)
-end)
+	RegisterNetEvent('esx:setWeaponTint')
+	AddEventHandler('esx:setWeaponTint', function(weapon, weaponTintIndex)
+		SetPedWeaponTintIndex(ESX.PlayerData.ped, GetHashKey(weapon), weaponTintIndex)
+	end)
+
+	RegisterNetEvent('esx:removeWeapon')
+	AddEventHandler('esx:removeWeapon', function(weapon)
+		local playerPed = ESX.PlayerData.ped
+		RemoveWeaponFromPed(ESX.PlayerData.ped, GetHashKey(weapon))
+		SetPedAmmo(ESX.PlayerData.ped, GetHashKey(weapon), 0)
+	end)
+
+	RegisterNetEvent('esx:removeWeaponComponent')
+	AddEventHandler('esx:removeWeaponComponent', function(weapon, weaponComponent)
+		local componentHash = ESX.GetWeaponComponent(weapon, weaponComponent).hash
+		RemoveWeaponComponentFromPed(ESX.PlayerData.ped, GetHashKey(weapon), componentHash)
+	end)
+end
 
 RegisterNetEvent('esx:teleport')
 AddEventHandler('esx:teleport', function(coords)
@@ -256,45 +263,47 @@ AddEventHandler('esx:spawnVehicle', function(vehicle)
 	end)
 end)
 
-RegisterNetEvent('esx:createPickup')
-AddEventHandler('esx:createPickup', function(pickupId, label, coords, type, name, components, tintIndex)
-	local function setObjectProperties(object)
-		SetEntityAsMissionEntity(object, true, false)
-		PlaceObjectOnGroundProperly(object)
-		FreezeEntityPosition(object, true)
-		SetEntityCollision(object, false, true)
+if not Config.OxInventory then
+	RegisterNetEvent('esx:createPickup')
+	AddEventHandler('esx:createPickup', function(pickupId, label, coords, type, name, components, tintIndex)
+		local function setObjectProperties(object)
+			SetEntityAsMissionEntity(object, true, false)
+			PlaceObjectOnGroundProperly(object)
+			FreezeEntityPosition(object, true)
+			SetEntityCollision(object, false, true)
 
-		pickups[pickupId] = {
-			obj = object,
-			label = label,
-			inRange = false,
-			coords = vector3(coords.x, coords.y, coords.z)
-		}
-	end
-
-	if type == 'item_weapon' then
-		local weaponHash = GetHashKey(name)
-		ESX.Streaming.RequestWeaponAsset(weaponHash)
-		local pickupObject = CreateWeaponObject(weaponHash, 50, coords.x, coords.y, coords.z, true, 1.0, 0)
-		SetWeaponObjectTintIndex(pickupObject, tintIndex)
-
-		for k,v in ipairs(components) do
-			local component = ESX.GetWeaponComponent(name, v)
-			GiveWeaponComponentToWeaponObject(pickupObject, component.hash)
+			pickups[pickupId] = {
+				obj = object,
+				label = label,
+				inRange = false,
+				coords = vector3(coords.x, coords.y, coords.z)
+			}
 		end
 
-		setObjectProperties(pickupObject)
-	else
-		ESX.Game.SpawnLocalObject('prop_money_bag_01', coords, setObjectProperties)
-	end
-end)
+		if type == 'item_weapon' then
+			local weaponHash = GetHashKey(name)
+			ESX.Streaming.RequestWeaponAsset(weaponHash)
+			local pickupObject = CreateWeaponObject(weaponHash, 50, coords.x, coords.y, coords.z, true, 1.0, 0)
+			SetWeaponObjectTintIndex(pickupObject, tintIndex)
 
-RegisterNetEvent('esx:createMissingPickups')
-AddEventHandler('esx:createMissingPickups', function(missingPickups)
-	for pickupId, pickup in pairs(missingPickups) do
-		TriggerEvent('esx:createPickup', pickupId, pickup.label, pickup.coords, pickup.type, pickup.name, pickup.components, pickup.tintIndex)
-	end
-end)
+			for k,v in ipairs(components) do
+				local component = ESX.GetWeaponComponent(name, v)
+				GiveWeaponComponentToWeaponObject(pickupObject, component.hash)
+			end
+
+			setObjectProperties(pickupObject)
+		else
+			ESX.Game.SpawnLocalObject('prop_money_bag_01', coords, setObjectProperties)
+		end
+	end)
+
+	RegisterNetEvent('esx:createMissingPickups')
+	AddEventHandler('esx:createMissingPickups', function(missingPickups)
+		for pickupId, pickup in pairs(missingPickups) do
+			TriggerEvent('esx:createPickup', pickupId, pickup.label, pickup.coords, pickup.type, pickup.name, pickup.components, pickup.tintIndex)
+		end
+	end)
+end
 
 RegisterNetEvent('esx:registerSuggestions')
 AddEventHandler('esx:registerSuggestions', function(registeredCommands)
@@ -305,13 +314,15 @@ AddEventHandler('esx:registerSuggestions', function(registeredCommands)
 	end
 end)
 
-RegisterNetEvent('esx:removePickup')
-AddEventHandler('esx:removePickup', function(pickupId)
-	if pickups[pickupId] and pickups[pickupId].obj then
-		ESX.Game.DeleteObject(pickups[pickupId].obj)
-		pickups[pickupId] = nil
-	end
-end)
+if not Config.OxInventory then
+	RegisterNetEvent('esx:removePickup')
+	AddEventHandler('esx:removePickup', function(pickupId)
+		if pickups[pickupId] and pickups[pickupId].obj then
+			ESX.Game.DeleteObject(pickups[pickupId].obj)
+			pickups[pickupId] = nil
+		end
+	end)
+end
 
 RegisterNetEvent('esx:deleteVehicle')
 AddEventHandler('esx:deleteVehicle', function(radius)
@@ -323,7 +334,7 @@ AddEventHandler('esx:deleteVehicle', function(radius)
 			local attempt = 0
 
 			while not NetworkHasControlOfEntity(entity) and attempt < 100 and DoesEntityExist(entity) do
-				Citizen.Wait(100)
+				Wait(100)
 				NetworkRequestControlOfEntity(entity)
 				attempt = attempt + 1
 			end
@@ -340,7 +351,7 @@ AddEventHandler('esx:deleteVehicle', function(radius)
 		end
 
 		while not NetworkHasControlOfEntity(vehicle) and attempt < 100 and DoesEntityExist(vehicle) do
-			Citizen.Wait(100)
+			Wait(100)
 			NetworkRequestControlOfEntity(vehicle)
 			attempt = attempt + 1
 		end
@@ -353,11 +364,11 @@ end)
 
 -- Pause menu disables HUD display
 if Config.EnableHud then
-	Citizen.CreateThread(function()
+	CreateThread(function()
 		local isPaused = false
 		local time = 500
 		while true do
-			Citizen.Wait(time)
+			Wait(time)
 
 			if IsPauseMenuActive() and not isPaused then
 				time = 100
@@ -377,40 +388,42 @@ if Config.EnableHud then
 end
 
 function StartServerSyncLoops()
-	-- keep track of ammo
-	Citizen.CreateThread(function()
-		local currentWeapon = {timer=0}
-		while ESX.PlayerLoaded do
-			local sleep = 5
+	if not Config.OxInventory then
+		-- keep track of ammo
+		CreateThread(function()
+			local currentWeapon = {timer=0}
+			while ESX.PlayerLoaded do
+				local sleep = 5
 
-			if currentWeapon.timer == sleep then
-				local ammoCount = GetAmmoInPedWeapon(ESX.PlayerData.ped, currentWeapon.hash)
-				TriggerServerEvent('esx:updateWeaponAmmo', currentWeapon.name, ammoCount)
-				currentWeapon.timer = 0
-			elseif currentWeapon.timer > sleep then
-				currentWeapon.timer = currentWeapon.timer - sleep
-			end
-
-			if IsPedArmed(ESX.PlayerData.ped, 4) then
-				if IsPedShooting(ESX.PlayerData.ped) then
-					local _,weaponHash = GetCurrentPedWeapon(ESX.PlayerData.ped, true)
-					local weapon = ESX.GetWeaponFromHash(weaponHash)
-
-					if weapon then
-						currentWeapon.name = weapon.name
-						currentWeapon.hash = weaponHash	
-						currentWeapon.timer = 100 * sleep		
-					end
+				if currentWeapon.timer == sleep then
+					local ammoCount = GetAmmoInPedWeapon(ESX.PlayerData.ped, currentWeapon.hash)
+					TriggerServerEvent('esx:updateWeaponAmmo', currentWeapon.name, ammoCount)
+					currentWeapon.timer = 0
+				elseif currentWeapon.timer > sleep then
+					currentWeapon.timer = currentWeapon.timer - sleep
 				end
-			else
-				sleep = 200
+
+				if IsPedArmed(ESX.PlayerData.ped, 4) then
+					if IsPedShooting(ESX.PlayerData.ped) then
+						local _,weaponHash = GetCurrentPedWeapon(ESX.PlayerData.ped, true)
+						local weapon = ESX.GetWeaponFromHash(weaponHash)
+
+						if weapon then
+							currentWeapon.name = weapon.name
+							currentWeapon.hash = weaponHash
+							currentWeapon.timer = 100 * sleep
+						end
+					end
+				else
+					sleep = 200
+				end
+				Wait(sleep)
 			end
-			Citizen.Wait(sleep)
-		end
-	end)
+		end)
+	end
 
 	-- sync current player coords with server
-	Citizen.CreateThread(function()
+	CreateThread(function()
 		local previousCoords = vector3(ESX.PlayerData.coords.x, ESX.PlayerData.coords.y, ESX.PlayerData.coords.z)
 
 		while ESX.PlayerLoaded do
@@ -428,12 +441,12 @@ function StartServerSyncLoops()
 					TriggerServerEvent('esx:updateCoords', formattedCoords)
 				end
 			end
-			Citizen.Wait(1500)
+			Wait(1500)
 		end
 	end)
 end
 
-if Config.EnableDefaultInventory then
+if not Config.OxInventory and Config.EnableDefaultInventory then
 	RegisterCommand('showinv', function()
 		if not ESX.PlayerData.dead and not ESX.UI.Menu.IsOpen('default', 'es_extended', 'inventory') then
 			ESX.ShowInventory()
@@ -449,52 +462,54 @@ if not Config.EnableWantedLevel then
 	SetMaxWantedLevel(0)
 end
 
-Citizen.CreateThread(function()
-	while true do
-		Citizen.Wait(0)
-		local playerCoords, letSleep = GetEntityCoords(ESX.PlayerData.ped), true
-		local closestPlayer, closestDistance = ESX.Game.GetClosestPlayer(playerCoords)
+if not Config.OxInventory then
+	CreateThread(function()
+		while true do
+			Wait(0)
+			local playerCoords, letSleep = GetEntityCoords(ESX.PlayerData.ped), true
+			local closestPlayer, closestDistance = ESX.Game.GetClosestPlayer(playerCoords)
 
-		for pickupId,pickup in pairs(pickups) do
-			local distance = #(playerCoords - pickup.coords)
+			for pickupId,pickup in pairs(pickups) do
+				local distance = #(playerCoords - pickup.coords)
 
-			if distance < 5 then
-				local label = pickup.label
-				letSleep = false
+				if distance < 5 then
+					local label = pickup.label
+					letSleep = false
 
-				if distance < 1 then
-					if IsControlJustReleased(0, 38) then
-						if IsPedOnFoot(ESX.PlayerData.ped) and (closestDistance == -1 or closestDistance > 3) and not pickup.inRange then
-							pickup.inRange = true
+					if distance < 1 then
+						if IsControlJustReleased(0, 38) then
+							if IsPedOnFoot(ESX.PlayerData.ped) and (closestDistance == -1 or closestDistance > 3) and not pickup.inRange then
+								pickup.inRange = true
 
-							local dict, anim = 'weapons@first_person@aim_rng@generic@projectile@sticky_bomb@', 'plant_floor'
-							ESX.Streaming.RequestAnimDict(dict)
-							TaskPlayAnim(ESX.PlayerData.ped, dict, anim, 8.0, 1.0, 1000, 16, 0.0, false, false, false)
-							Citizen.Wait(1000)
+								local dict, anim = 'weapons@first_person@aim_rng@generic@projectile@sticky_bomb@', 'plant_floor'
+								ESX.Streaming.RequestAnimDict(dict)
+								TaskPlayAnim(ESX.PlayerData.ped, dict, anim, 8.0, 1.0, 1000, 16, 0.0, false, false, false)
+								Wait(1000)
 
-							TriggerServerEvent('esx:onPickup', pickupId)
-							PlaySoundFrontend(-1, 'PICK_UP', 'HUD_FRONTEND_DEFAULT_SOUNDSET', false)
+								TriggerServerEvent('esx:onPickup', pickupId)
+								PlaySoundFrontend(-1, 'PICK_UP', 'HUD_FRONTEND_DEFAULT_SOUNDSET', false)
+							end
 						end
+
+						label = ('%s~n~%s'):format(label, _U('threw_pickup_prompt'))
 					end
 
-					label = ('%s~n~%s'):format(label, _U('threw_pickup_prompt'))
+					ESX.Game.Utils.DrawText3D({
+						x = pickup.coords.x,
+						y = pickup.coords.y,
+						z = pickup.coords.z + 0.25
+					}, label, 1.2, 1)
+				elseif pickup.inRange then
+					pickup.inRange = false
 				end
+			end
 
-				ESX.Game.Utils.DrawText3D({
-					x = pickup.coords.x,
-					y = pickup.coords.y,
-					z = pickup.coords.z + 0.25
-				}, label, 1.2, 1)
-			elseif pickup.inRange then
-				pickup.inRange = false
+			if letSleep then
+				Wait(500)
 			end
 		end
-
-		if letSleep then
-			Citizen.Wait(500)
-		end
-	end
-end)
+	end)
+end
 
 ----- Admin commnads from esx_adminplus
 
@@ -517,7 +532,7 @@ AddEventHandler("esx:tpm", function()
                 break
             	end
 
-            Citizen.Wait(5)
+            Wait(0)
         	end
         TriggerEvent('chatMessage', "Successfully Teleported")
     	else
@@ -533,7 +548,7 @@ AddEventHandler("esx:noclip", function(input)
 	ESX.TriggerServerCallback("esx:isUserAdmin", function(admin)
 		if admin then
     local player = PlayerId()
-	
+
     local msg = "disabled"
 	if(noclip == false)then
 		noclip_pos = GetEntityCoords(ESX.PlayerData.ped, false)
@@ -549,11 +564,11 @@ AddEventHandler("esx:noclip", function(input)
 	end
 	end)
 end)
-	
+
 	local heading = 0
-	Citizen.CreateThread(function()
+	CreateThread(function()
 	while true do
-		Citizen.Wait(0)
+		Wait(0)
 
 		if(noclip)then
 			SetEntityCoordsNoOffset(ESX.PlayerData.ped, noclip_pos.x, noclip_pos.y, noclip_pos.z, 0, 0, 0)
@@ -592,7 +607,7 @@ end)
 				noclip_pos = GetOffsetFromEntityInWorldCoords(ESX.PlayerData.ped, 0.0, 0.0, -1.0)
 			end
 		else
-			Citizen.Wait(200)
+			Wait(200)
 		end
 	end
 end)
